// src/models/agentrunsbatchresponse.ts
/*
 * Code generated by Speakeasy (https://speakeasy.com). DO NOT EDIT.
 */

import * as z from "zod/v4-mini";
import { safeParse } from "../lib/schemas.js";
import { ClosedEnum } from "../types/enums.js";
import { Result as SafeParseResult } from "../types/fp.js";
import * as types from "../types/primitives.js";
import {
  AgentRunsResponseOutput,
  AgentRunsResponseOutput$inboundSchema,
} from "./agentrunsresponseoutput.js";
import { SDKValidationError } from "./errors/sdkvalidationerror.js";

/**
 * The access based role of the user
 */
export const Role = {
  User: "user",
} as const;
/**
 * The access based role of the user
 */
export type Role = ClosedEnum<typeof Role>;

export type Input = {
  /**
   * The access based role of the user
   */
  role: Role;
  /**
   * The question populated in the request payload
   */
  content: string;
};

export type AgentRunsBatchResponse = {
  /**
   * The id of the agent populated in the request.
   */
  agent: string;
  /**
   * The mode of the agent
   */
  mode?: string | undefined;
  /**
   * The users access role and question you asked the agent
   */
  input: Array<Input>;
  /**
   * Array of response outputs from the agent
   */
  output: Array<AgentRunsResponseOutput>;
};

/** @internal */
export const Role$inboundSchema: z.ZodMiniEnum<typeof Role> = z.enum(Role);

/** @internal */
export const Input$inboundSchema: z.ZodMiniType<Input, unknown> = z.object({
  role: Role$inboundSchema,
  content: types.string(),
});

export function inputFromJSON(
  jsonString: string,
): SafeParseResult<Input, SDKValidationError> {
  return safeParse(
    jsonString,
    (x) => Input$inboundSchema.parse(JSON.parse(x)),
    `Failed to parse 'Input' from JSON`,
  );
}

/** @internal */
export const AgentRunsBatchResponse$inboundSchema: z.ZodMiniType<
  AgentRunsBatchResponse,
  unknown
> = z.object({
  agent: types.string(),
  mode: types.optional(types.string()),
  input: z.array(z.lazy(() => Input$inboundSchema)),
  output: z.array(AgentRunsResponseOutput$inboundSchema),
});

export function agentRunsBatchResponseFromJSON(
  jsonString: string,
): SafeParseResult<AgentRunsBatchResponse, SDKValidationError> {
  return safeParse(
    jsonString,
    (x) => AgentRunsBatchResponse$inboundSchema.parse(JSON.parse(x)),
    `Failed to parse 'AgentRunsBatchResponse' from JSON`,
  );
}